TreeCountAndHealth/FastAPIapp_UploadedToCloudRun/requirements.txt
torch==2.2.2+cpu
torchvision==0.17.2+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html


fastapi
uvicorn
ultralytics
numpy<2
opencv-python-headless
python-multipart
google-cloud-storage
google-cloud-firestore